Actualización luego de pruebas en RoomController.

+++ G6.ProBook.WebApi/DTOs/RoomDto.cs
@@ -9,7 +9,7 @@
         public string Type { get; set; } = string.Empty;
         public int Capacity { get; set; }
         public List<string> Amenities { get; set; } = new List<string>();
-        public decimal BasePrice { get; set; }
+        public double BasePrice { get; set; }
         public string PhotoUrl { get; set; } = string.Empty;
     }
 }
